langchain/requirements.txt
langchain
langchain_community
langgraph
langserve
python-dotenv
# langchain_google.genai
langchain-google-genai
fastapi
uvicorn
sse_starlette